Add files via upload
The trex file is a simulation model structure for this project that was programmed in TreeAge Pro Healthcare software. A cohort-level Markov approach was used. The png file is a screenshot of model structure from TreeAge software. The R file are codes for further analysis of the outputs that were exported from TreeAge and imported to R, mainly for better visualisation.